+++ saleor/app/models.py
@@ -11,7 +11,12 @@ from ..core.models import Job, ModelWithMetadata
 from ..permission.enums import AppPermission, BasePermissionEnum
 from ..permission.models import Permission
 from ..webhook.event_types import WebhookEventAsyncType, WebhookEventSyncType
-from .types import AppExtensionMount, AppExtensionTarget, AppType
+from .types import (
+    AppExtensionHttpMethod,
+    AppExtensionMount,
+    AppExtensionTarget,
+    AppType,
+)
 
 
 class AppQueryset(models.QuerySet["App"]):
@@ -162,6 +167,11 @@ class AppExtension(models.Model):
         Permission,
         blank=True,
         help_text="Specific permissions for this app extension.",
+    )
+    http_target_method = models.CharField(
+        blank=False,
+        null=True,
+        choices=AppExtensionHttpMethod.CHOICES,
     )