+++ saleor/app/types.py
@@ -15,18 +15,22 @@ class AppExtensionMount:
     COLLECTION_OVERVIEW_CREATE = "collection_overview_create"
     COLLECTION_OVERVIEW_MORE_ACTIONS = "collection_overview_more_actions"
     COLLECTION_DETAILS_MORE_ACTIONS = "collection_details_more_actions"
+    COLLECTION_DETAILS_WIDGETS = "collection_details_widgets"
 
     GIFT_CARD_OVERVIEW_CREATE = "gift_card_overview_create"
     GIFT_CARD_OVERVIEW_MORE_ACTIONS = "gift_card_overview_more_actions"
     GIFT_CARD_DETAILS_MORE_ACTIONS = "gift_card_details_more_actions"
+    GIFT_CARD_DETAILS_WIDGETS = "gift_card_details_widgets"
 
     CUSTOMER_OVERVIEW_CREATE = "customer_overview_create"
     CUSTOMER_OVERVIEW_MORE_ACTIONS = "customer_overview_more_actions"
     CUSTOMER_DETAILS_MORE_ACTIONS = "customer_details_more_actions"
+    CUSTOMER_DETAILS_WIDGETS = "customer_details_widgets"
 
     PRODUCT_OVERVIEW_CREATE = "product_overview_create"
     PRODUCT_OVERVIEW_MORE_ACTIONS = "product_overview_more_actions"
     PRODUCT_DETAILS_MORE_ACTIONS = "product_details_more_actions"
+    PRODUCT_DETAILS_WIDGETS = "product_details_widgets"
 
     NAVIGATION_CATALOG = "navigation_catalog"
     NAVIGATION_ORDERS = "navigation_orders"
@@ -38,10 +42,12 @@ class AppExtensionMount:
     ORDER_DETAILS_MORE_ACTIONS = "order_details_more_actions"
     ORDER_OVERVIEW_CREATE = "order_overview_create"
     ORDER_OVERVIEW_MORE_ACTIONS = "order_overview_more_actions"
+    ORDER_DETAILS_WIDGETS = "order_details_widgets"
 
     DRAFT_ORDER_DETAILS_MORE_ACTIONS = "draft_order_details_more_actions"
     DRAFT_ORDER_OVERVIEW_CREATE = "draft_order_overview_create"
     DRAFT_ORDER_OVERVIEW_MORE_ACTIONS = "draft_order_overview_more_actions"
+    DRAFT_ORDER_DETAILS_WIDGETS = "draft_order_details_widgets"
 
     DISCOUNT_DETAILS_MORE_ACTIONS = "discount_details_more_actions"
     DISCOUNT_OVERVIEW_CREATE = "discount_overview_create"
@@ -50,6 +56,7 @@ class AppExtensionMount:
     VOUCHER_DETAILS_MORE_ACTIONS = "voucher_details_more_actions"
     VOUCHER_OVERVIEW_CREATE = "voucher_overview_create"
     VOUCHER_OVERVIEW_MORE_ACTIONS = "voucher_overview_more_actions"
+    VOUCHER_DETAILS_WIDGETS = "voucher_details_widgets"
 
     PAGE_DETAILS_MORE_ACTIONS = "page_details_more_actions"
     PAGE_OVERVIEW_CREATE = "page_overview_create"
@@ -70,15 +77,19 @@ class AppExtensionMount:
         (COLLECTION_OVERVIEW_CREATE, "collection_overview_create"),
         (COLLECTION_OVERVIEW_MORE_ACTIONS, "collection_overview_more_actions"),
         (COLLECTION_DETAILS_MORE_ACTIONS, "collection_details_more_actions"),
+        (COLLECTION_DETAILS_WIDGETS, "collection_details_widgets"),
         (GIFT_CARD_OVERVIEW_CREATE, "gift_card_overview_create"),
         (GIFT_CARD_OVERVIEW_MORE_ACTIONS, "gift_card_overview_more_actions"),
         (GIFT_CARD_DETAILS_MORE_ACTIONS, "gift_card_details_more_actions"),
+        (GIFT_CARD_DETAILS_WIDGETS, "gift_card_details_widgets"),
         (CUSTOMER_OVERVIEW_CREATE, "customer_overview_create"),
         (CUSTOMER_OVERVIEW_MORE_ACTIONS, "customer_overview_more_actions"),
         (CUSTOMER_DETAILS_MORE_ACTIONS, "customer_details_more_actions"),
+        (CUSTOMER_DETAILS_WIDGETS, "customer_details_widgets"),
         (PRODUCT_OVERVIEW_CREATE, "product_overview_create"),
         (PRODUCT_OVERVIEW_MORE_ACTIONS, "product_overview_more_actions"),
         (PRODUCT_DETAILS_MORE_ACTIONS, "product_details_more_actions"),
+        (PRODUCT_DETAILS_WIDGETS, "product_details_widgets"),
         (NAVIGATION_CATALOG, "navigation_catalog"),
         (NAVIGATION_ORDERS, "navigation_orders"),
         (NAVIGATION_CUSTOMERS, "navigation_customers"),
@@ -88,15 +99,18 @@ class AppExtensionMount:
         (ORDER_DETAILS_MORE_ACTIONS, "order_details_more_actions"),
         (ORDER_OVERVIEW_CREATE, "order_overview_create"),
         (ORDER_OVERVIEW_MORE_ACTIONS, "order_overview_more_actions"),
+        (ORDER_DETAILS_WIDGETS, "order_details_widgets"),
         (DRAFT_ORDER_DETAILS_MORE_ACTIONS, "draft_order_details_more_actions"),
         (DRAFT_ORDER_OVERVIEW_CREATE, "draft_order_overview_create"),
         (DRAFT_ORDER_OVERVIEW_MORE_ACTIONS, "draft_order_overview_more_actions"),
+        (DRAFT_ORDER_DETAILS_WIDGETS, "draft_order_details_widgets"),
         (DISCOUNT_DETAILS_MORE_ACTIONS, "discount_details_more_actions"),
         (DISCOUNT_OVERVIEW_CREATE, "discount_overview_create"),
         (DISCOUNT_OVERVIEW_MORE_ACTIONS, "discount_overview_more_actions"),
         (VOUCHER_DETAILS_MORE_ACTIONS, "voucher_details_more_actions"),
         (VOUCHER_OVERVIEW_CREATE, "voucher_overview_create"),
         (VOUCHER_OVERVIEW_MORE_ACTIONS, "voucher_overview_more_actions"),
+        (VOUCHER_DETAILS_WIDGETS, "voucher_details_widgets"),
         (PAGE_DETAILS_MORE_ACTIONS, "page_details_more_actions"),
         (PAGE_OVERVIEW_CREATE, "page_overview_create"),
         (PAGE_OVERVIEW_MORE_ACTIONS, "page_overview_more_actions"),
@@ -119,5 +133,23 @@ class AppExtensionTarget:
     POPUP = "popup"
     APP_PAGE = "app_page"
     NEW_TAB = "new_tab"
+    WIDGET = "widget"
 
-    CHOICES = [(POPUP, "popup"), (APP_PAGE, "app_page"), (NEW_TAB, "new_tab")]
+    CHOICES = [
+        (POPUP, "popup"),
+        (APP_PAGE, "app_page"),
+        (NEW_TAB, "new_tab"),
+        (WIDGET, "widget"),
+    ]
+
+
+class AppExtensionHttpMethod:
+    """HTTP methods available for app extensions.
+
+    Represents available HTTPS methods for frontend to work with extension (WIDGET and NEW_TAB)
+    """
+
+    GET = "GET"
+    POST = "POST"
+
+    CHOICES = [("GET", "GET"), ("POST", "POST")]