+++ saleor/graphql/app/types.py
@@ -4,7 +4,7 @@ import datetime
 import graphene
 
 from ...app import models
-from ...app.types import AppExtensionTarget
+from ...app.types import AppExtensionHttpMethod, AppExtensionTarget
 from ...core.exceptions import PermissionDenied
 from ...core.jwt import JWT_THIRDPARTY_ACCESS_TYPE
 from ...core.utils import build_absolute_uri
@@ -27,11 +27,12 @@ from ..core import ResolveInfo, SaleorContext
 from ..core.connection import CountableConnection
 from ..core.context import get_database_connection_name
 from ..core.dataloaders import DataLoader
-from ..core.descriptions import ADDED_IN_319, ADDED_IN_321
+from ..core.descriptions import ADDED_IN_319, ADDED_IN_321, ADDED_IN_322
 from ..core.doc_category import DOC_CATEGORY_APPS
 from ..core.federation import federated_entity, resolve_federation_references
 from ..core.scalars import DateTime
 from ..core.types import (
+    BaseEnum,
     BaseObjectType,
     IconThumbnailField,
     Job,
@@ -137,6 +138,64 @@ class AppManifestExtension(BaseObjectType):
         return resolve_app_extension_url(root)
 
 
+class HttpMethod(BaseEnum):
+    POST = AppExtensionHttpMethod.POST
+    GET = AppExtensionHttpMethod.GET
+
+
+class NewTabTargetOptions(BaseObjectType):
+    method = graphene.Field(
+        HttpMethod,
+        required=True,
+        description="HTTP method for New Tab target (GET or POST)",
+    )
+
+    class Meta:
+        description = "Represents the NEW_TAB target options for an app extension."
+        doc_category = DOC_CATEGORY_APPS
+
+
+class WidgetTargetOptions(BaseObjectType):
+    method = graphene.Field(
+        HttpMethod,
+        required=True,
+        description="HTTP method for Widget target (GET or POST)",
+    )
+
+    class Meta:
+        description = "Represents the WIDGET target options for an app extension."
+        doc_category = DOC_CATEGORY_APPS
+
+
+class AppExtensionOptionsWidget(BaseObjectType):
+    widget_target = graphene.Field(
+        WidgetTargetOptions,
+        description="Options for displaying a Widget",
+        required=False,
+    )
+
+    class Meta:
+        description = "Represents the options for an app extension."
+        doc_category = DOC_CATEGORY_APPS
+
+
+class AppExtensionOptionsNewTab(BaseObjectType):
+    new_tab_target = graphene.Field(
+        NewTabTargetOptions,
+        description="Options controlling behavior of the NEW_TAB extension target",
+        required=False,
+    )
+
+    class Meta:
+        description = "Represents the options for an app extension."
+        doc_category = DOC_CATEGORY_APPS
+
+
+class AppExtensionPossibleOptions(graphene.Union):
+    class Meta:
+        types = (AppExtensionOptionsWidget, AppExtensionOptionsNewTab)
+
+
 class AppExtension(AppManifestExtension, ModelObjectType[models.AppExtension]):
     id = graphene.GlobalID(required=True, description="The ID of the app extension.")
     app = graphene.Field(
@@ -146,6 +205,10 @@ class AppExtension(AppManifestExtension, ModelObjectType[models.AppExtension]):
     )
     access_token = graphene.String(
         description="JWT token used to authenticate by third-party app extension."
+    )
+    options = graphene.Field(
+        AppExtensionPossibleOptions,
+        description="App extension options." + ADDED_IN_322,
     )
 
     class Meta:
@@ -200,6 +263,22 @@ class AppExtension(AppManifestExtension, ModelObjectType[models.AppExtension]):
             return resolve_access_token_for_app_extension(info, root, app)
 
         return AppByIdLoader(info.context).load(root.app_id).then(_resolve_access_token)
+
+    @staticmethod
+    def resolve_options(root: models.AppExtension, _info: ResolveInfo):
+        http_method = root.http_target_method
+
+        if root.target == AppExtensionTarget.WIDGET:
+            return AppExtensionOptionsWidget(
+                widget_target=WidgetTargetOptions(method=http_method),
+            )
+
+        if root.target == AppExtensionTarget.NEW_TAB:
+            return AppExtensionOptionsNewTab(
+                new_tab_target=NewTabTargetOptions(method=http_method),
+            )
+
+        return None
 
 
 class AppExtensionCountableConnection(CountableConnection):